Revert "🐛 Cast micros() into double"
This reverts commit 8ddec5fa9893e5d2cd1d824d358c7611c42db9fb.

+++ homework5/vrduino/OrientationTracker.cpp
@@ -191,7 +191,7 @@ bool OrientationTracker::updateImuVariables() {
   //update:
   //previousTimeImu (in seconds)
   //deltaT (in seconds)
-  double currentTimeImu = micros() / 1000000;
+  double currentTimeImu = ((double) micros()) / 1000000;
   deltaT = currentTimeImu - previousTimeImu;
   previousTimeImu = currentTimeImu;
